requirements.txt
APScheduler==3.9.1
attrs==20.3.0
bcrypt==3.2.2
blinker==1.4
cachelib==0.7.0
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
click==8.1.3
cloud-init==20.4.1
configobj==5.0.6
cryptography==3.3.2
dictdiffer==0.9.0
dnspython==2.2.1
email-validator==1.2.1
Flask==2.1.2
Flask-Bcrypt==1.0.1
flask-mongoengine==1.0.0
Flask-Session==0.4.0
Flask-WTF==1.0.1
gunicorn==20.1.0
httplib2==0.18.1
idna==2.10
importlib-metadata==4.11.4
iotop==0.6
itsdangerous==2.1.2
Jinja2==3.1.2
jsonpatch==1.25
jsonpointer==2.0
jsonschema==3.2.0
MarkupSafe==2.1.1
mongoengine==0.24.1
more-itertools==4.2.0
oauthlib==3.1.0
pycparser==2.21
pycryptodome==3.14.1
pycurl==7.43.0.6
PyJWT==1.7.1
pyminizip==0.2.6
pymongo==4.1.1
pyrsistent==0.15.5
PySimpleSOAP==1.16.2
python-apt==2.2.1
python-debian==0.1.39
python-debianbts==3.1.0
pytz==2022.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==5.3.1
reportbug==7.10.3+deb11u1
requests==2.25.1
six==1.16.0
tzdata==2022.1
tzlocal==4.2
ufw==0.36
urllib3==1.26.5
Werkzeug==2.1.2
WTForms==3.0.1
zipp==1.0.0
